Fix active courses listing for production site
Apparently SQLite and Postgres treat NULL values differently, which is
why the previous version of the page worked locally but doesn't work on
the beta. This commit should fix it. The | bitwise operator on two
QuerySets works to merge them.

Actually, I'm not completely sure this will work, because the order may
not be preserved. Hopefully it does. Last attempt for now.

I also changed Course.objects.all to Course.objects because there's no
need to evaluate the QuerySet in this case.