Convert tabs to spaces (4-space)
I knew this day would come!

For future reference, the command used was:

find ./ -name "*.py" -or -name "*.html" -or -name "*.js" -or -name
    "*.less" | xargs expand_spaces

where expand_spaces is a shell script that just does

    expand -t 4 < $f > /tmp/lol && cat /tmp/lol > $f

for every $f in $@.

I had to make some manual changes in wiki/tests/markdown.py to account
for the leading " on some lines (expand took that to mean the string
should start with 3 spaces instead of 4, which led to undesired
results). Also fixed an unaligned colon.