Show series pages on course overview page
Series are now first-class citizens, and can be hidden if necessary (so
I can finally recreate the series for ODEs). Pages that are not part of
a series have been relegated to second-class citizens, and show up on
the lower half of the page under "Miscellaneous pages". This makes it
easier to manage lecture note orderings for pages that have not been
created in the correct order, for instance.

This also removes the need for the series view.

I also added a bit of js for collapsing series/page tables.