Increase bottom margin for breadcrumbs
The amount of space above and below seem to match better now. Not a
perfect match by any means, but it's better. For some reason setting
them to the same amount of space actually makes it look worse.

Also, not gonna lie, I only made this 100% frivolous commit because
my previous one was unexpectedly 28 seconds short of 3 AM and so Github
counted it for the previous day. This contribution streak is starting
to feel like more trouble to maintain than it's worth.