Add id to each row of commit history table.
Add an id to each row of the commit history table to allow internal links to the appropriate commit. The purpose of this is to allow linking directly to the changeset of a page from the dashboard and other pages. Though you could link to the commit page, this will allow for more context to be kept while reviewing changes to courses.