Convert get_metadata_fields() into attribute
For PageType. There's no point in it being a method. In fact, it should
eventually be removed completely, since it's redundant (see the
editable_fields attribute), but for some reason I'm using professor_id
instead of professor in editable_fields. Hopefully fixing that can be
the next commit.

EDIT: Never mind. Can't be fixed at the moment. Will leave for another
day.